pyproject.toml
[tool.poetry]
name = "comp"
version = "0.1.0"
description = ""
authors = ["Jared Coughlin <jcoughl3@illinois.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.5"
h5py = "^3.1.0"
PyYAML = "^5.3.1"
progress = "^1.5"
unyt = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.7.0"

[tool.poetry.scripts]
ytcomp = 'comp.run:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
